test(e2e): correctly handle errors in the `runTests()` function
Previously, if the `runTests()` function failed (i.e. it did not even
get to run the tests), the error was ignored and the command succeeded.

This commit fixes it by ensurring that such failures are logged and
cause the command to fail. Additionally, it prevents failure to clean
up the version directories to affect the end result of the tests.