cfg(package): run e2e tests on the min supported VSCode version by default
This works around a [VSCode limitation][1] that prevents e2e tests to be
run from the command-line if another VSCode instance of the same version
is already running.

Since (in the majority of cases) the min supported version is different
than the the version used for development (typically latest stable or
insiders version), this change allows e2e tests (and dependent npm
scripts) to be successfully run from the command-line during local
development.

[1]: https://code.visualstudio.com/api/working-with-extensions/testing-extension#using-insiders-version-for-extension-development