feat(fixGuideCodeSnippetsPlugin): add `markdown-it` plugin to show code snippets in guide previews
Code snippets are included in guides as `<code-example>` elements or
`<code-tabs>` with one or more `<code-pane>` elements (for tabbed code
snippets). In guide previews, these unknown HTML elements are not shown
at all, making the preview a bit misleading.

This commit fixes this by adding a new `markdown-it` plugin that
processes these elements to show them in the previews as code blocks
(with appropriate formatting to improve readability).