fix(fixGuideImagesPlugin): rewrite image URLs to their original location
Previously, image production URLs in guide Markdown previews were
rewritten to their on-disk paths at `aio/src/generated/images/`.
However, this was not the images' original on-disk location. They are
copied there during development from `aio/content/images/`.

As a result, it is not guaranteed that the images exist in
`aio/src/generated/images/` (e.g. if the developer has not run the
script that prepares docs assets for the build).

This commit fixes the problem by rewritting the image URLs to point to
the original on-disk location, where they are guaranteed to work and
thus correctly show up in the preview.