perf(*): reduce size of bundled extension (no Webpack)
The size is reduced by excluding unused images from the extension
bundle.

---
This is mainly a commit to log my experience from investigating
switching to Webpack for bundling (which I decided against of).

Based on the [Bundling Extension][1] guide, I investigated switching to
Webpack for bundling. This proved problematic for a number of reasons:
- Producing the different compilations needed for various purposes
  (publishing the extension, running the extension locally, running unit
  tests, running e2e tests, etc.) proved quite complex. It would require
  lots of boilerplate config code, higher maintenance and/or use of
  proprietary Webpack features (such as `require.context()`).
- All examples I could find either did not include tests or used
  different tooling for bundling the extension for tests vs for
  publishing. This is problematic, since the published bundle will not
  be verified by tests.
- It is likely the case that there _are_ ways to achieve what I want in
  a maintainable way, but the fact that I spent several hours on this
  and still could not figure it out is indicative of the high
  maintenance cost.

Taking the above into account, as well as the fact that the (compressed)
bundle size did not get any smaller with Webpack, I decided to keep the
current build process (at least for now).

[1]: https://code.visualstudio.com/api/working-with-extensions/bundling-extension