ci(*): do not test on Node.js v8 on CI
Node.js v8 is not an Active LTS version (it is Maintenance LTS) and will
be reaching Enf-of-life status in less than a couple of months
(according to the Node.js [Releases docs][1]).

Testing on the latest 3 Node.js version, namely v10 (Active LTS), v12
(Active LTS) and v13 (Current) should be more than enough.

Note that the version of Node.js is only relevant for working on the
project (e.g. running tests, building, packaging, etc.). The extension
itself is run inside the user's VSCode, which provides its own Node.js
runtime.

[1]: https://nodejs.org/en/about/releases