ci(Travis): test on Windows as well
This commit adds Windows to the platforms tested against on Travis CI.
This is a prerequisite for moving to a single CI provider and ensure
deployments only happen if all tests (on all platforms) pass.

This is part of an investigation. It might not work out.

NOTE:
Secrets are not automatically filtered when targeting Windows on
Travis CI ([known issue][1]). This is OK, because we are not using any
secrets in tests and deployments (which do use secrets) are not run from
Windows.

[1]: https://travis-ci.community/t/current-known-issues-please-read-this-before-posting-a-new-topic/264