fix(CodeSnippetUtils): detect code example elements with whitespace-only content between tags
Previously, code example elements where only detected as such if the
closing tag was on the same or next line as the end of the corresponding
opening tag. In some cases, however, the opening and closing tags are
separated by newlines.

Since it is inexpensive to detect whitespace-only lines, this commit
adds support for detecting code example elements even when the opening
and closing tags are separated by lines containing whitespace-only
content.