feat(*): convert the extension to a web extension
Make the extension compatible with VSCode for the web (i.e. the
in-browser version of VSCode, as seen on `github.dev` for example).

I mostly followed the instructions in the [Web Extensions][1] guide.

NOTE:
Due to the complexity of running in-browser tests and the low impact
(given that there are no e2e tests), this commit does not add the
necessary infrastructure/configuration to run automated tests in a
browser environment.
For the time being, the extension can be manually tested using the
`Extension (web)` [launch configuration][2]. (Keep in mind that there is
currently an issue with contributed `markdown-it` plugins:
microsoft/vscode#133399)

[1]: https://code.visualstudio.com/api/extension-guides/web-extensions
[2]: https://code.visualstudio.com/api/extension-guides/web-extensions#test-your-web-extension-in-vs-code-running-on-desktop