fix(code-snippet-intellisense): use URIs instead of paths for locating example files
Previously, the code-snippet intellisense implementation relied on
file-system paths to resolve and locate example files. This worked fine
when using a local version of VSCode, but did not work when using the
extension on the [web version of VSCode][1].

This commit addresses the issue by using [Uri][2]`s instead of
file-system paths.

[1]: https://code.visualstudio.com/api/extension-guides/web-extensions
[2]: https://code.visualstudio.com/api/references/vscode-api#Uri